move filter controls to shared module
