+++ src/app/shared/shared.module.ts
@@ -4,15 +4,23 @@ import { CommonModule } from '@angular/common';
 import { AuthGuard, DeactivateGuard, UnauthGuard,
   ChartsModule, DatepickerModule, ImageModule, SelectModule, SpinnerModule } from 'ngx-prx-styleguide';
 
+import { CannedRangeComponent } from './filter/canned-range.component';
+import { DateRangeComponent } from './filter/date-range.component';
+
 @NgModule({
-  declarations: [],
+  declarations: [
+    CannedRangeComponent,
+    DateRangeComponent
+  ],
   exports: [
     CommonModule,
     ChartsModule,
     DatepickerModule,
     ImageModule,
     SelectModule,
-    SpinnerModule
+    SpinnerModule,
+    CannedRangeComponent,
+    DateRangeComponent
   ],
   imports: [
     CommonModule,
